Move checking of structure-options to a common location (so it will be picked up by all roslyn languages).
