+++ src/Features/Core/Portable/Structure/Syntax/AbstractBlockStructureProvider.cs
@@ -2,10 +2,8 @@
 
 using System;
 using System.Collections.Immutable;
-using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.ErrorReporting;
-using Microsoft.CodeAnalysis.Options;
 using Microsoft.CodeAnalysis.PooledObjects;
 using Roslyn.Utilities;
 
@@ -69,81 +67,12 @@ namespace Microsoft.CodeAnalysis.Structure
             BlockSpanCollector.CollectBlockSpans(
                 context.Document, syntaxRoot, _nodeProviderMap, _triviaProviderMap, spans, context.CancellationToken);
 
-            UpdateAndAddSpans(context, spans);
-
-            spans.Free();
-        }
-
-        internal static void UpdateAndAddSpans(BlockStructureContext context, ArrayBuilder<BlockSpan> spans)
-        {
-            var options = context.Document.Project.Solution.Workspace.Options;
-            var language = context.Document.Project.Language;
-
-            var showIndentGuidesForCodeLevelConstructs = options.GetOption(BlockStructureOptions.ShowBlockStructureGuidesForCodeLevelConstructs, language);
-            var showIndentGuidesForDeclarationLevelConstructs = options.GetOption(BlockStructureOptions.ShowBlockStructureGuidesForDeclarationLevelConstructs, language);
-            var showIndentGuidesForCommentsAndPreprocessorRegions = options.GetOption(BlockStructureOptions.ShowBlockStructureGuidesForCommentsAndPreprocessorRegions, language);
-            var showOutliningForCodeLevelConstructs = options.GetOption(BlockStructureOptions.ShowOutliningForCodeLevelConstructs, language);
-            var showOutliningForDeclarationLevelConstructs = options.GetOption(BlockStructureOptions.ShowOutliningForDeclarationLevelConstructs, language);
-            var showOutliningForCommentsAndPreprocessorRegions = options.GetOption(BlockStructureOptions.ShowOutliningForCommentsAndPreprocessorRegions, language);
-
             foreach (var span in spans)
             {
-                var updatedSpan = UpdateBlockSpan(span,
-                        showIndentGuidesForCodeLevelConstructs,
-                        showIndentGuidesForDeclarationLevelConstructs,
-                        showIndentGuidesForCommentsAndPreprocessorRegions,
-                        showOutliningForCodeLevelConstructs,
-                        showOutliningForDeclarationLevelConstructs,
-                        showOutliningForCommentsAndPreprocessorRegions);
-                context.AddBlockSpan(updatedSpan);
-            }
-        }
-
-        internal static BlockSpan UpdateBlockSpan(BlockSpan blockSpan,
-            bool showIndentGuidesForCodeLevelConstructs,
-            bool showIndentGuidesForDeclarationLevelConstructs,
-            bool showIndentGuidesForCommentsAndPreprocessorRegions,
-            bool showOutliningForCodeLevelConstructs,
-            bool showOutliningForDeclarationLevelConstructs,
-            bool showOutliningForCommentsAndPreprocessorRegions)
-        {
-            var type = blockSpan.Type;
-
-            var isTopLevel = BlockTypes.IsDeclarationLevelConstruct(type);
-            var isMemberLevel = BlockTypes.IsCodeLevelConstruct(type);
-            var isComment = BlockTypes.IsCommentOrPreprocessorRegion(type);
-
-            if (!showIndentGuidesForDeclarationLevelConstructs && isTopLevel)
-            {
-                type = BlockTypes.Nonstructural;
-            }
-            else if (!showIndentGuidesForCodeLevelConstructs && isMemberLevel)
-            {
-                type = BlockTypes.Nonstructural;
-            }
-            else if (!showIndentGuidesForCommentsAndPreprocessorRegions && isComment)
-            {
-                type = BlockTypes.Nonstructural;
+                context.AddBlockSpan(span);
             }
 
-            var isCollapsible = blockSpan.IsCollapsible;
-            if (isCollapsible)
-            {
-                if (!showOutliningForDeclarationLevelConstructs && isTopLevel)
-                {
-                    isCollapsible = false;
-                }
-                else if (!showOutliningForCodeLevelConstructs && isMemberLevel)
-                {
-                    isCollapsible = false;
-                }
-                else if (!showOutliningForCommentsAndPreprocessorRegions && isComment)
-                {
-                    isCollapsible = false;
-                }
-            }
-
-            return blockSpan.With(type: type, isCollapsible: isCollapsible);
+            spans.Free();
         }
     }
 }
